src/pyproject.toml
[tool.poetry]
name = "text-search"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.1"
pydantic-settings = "^2.4.0"
elasticsearch = "^8.15.0"
psycopg2-binary = "^2.9.9"
fastapi = "^0.114.2"
orjson = "^3.10.7"
sqlalchemy = "^2.0.34"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
aiohttp = "^3.10.5"
greenlet = "^3.1.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
